Avoid redundant problemMatcher list in tasks.json
I think it's good practice to omit these, even if VSCode generates them automatically. If done so, the defaults from vshaxe's task generation will be used, which means it's more future proof if the list of problem matchers there ever changes. It also makes for a more consise task definition.