fix use hldx on ime and custom font bug
In hldx, if I customize a font (for example, it is Korean, Chinese), then I need to set hasChar to force it to return true, because the mapping of characters is generated in real time and it is impossible to traverse. But if I force hasChar to return true, it will cause font.hasChar to then go down and hldx will send ctrl and delete key as charkey, causing the program to go wrong.